PhaseII/backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
uvicorn[standard]==0.32.0
pydantic==2.10.1
pydantic-settings==2.7.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
pytest==8.3.3
pytest-asyncio==0.25.0
httpx==0.27.2
python-dotenv==1.0.1
cryptography==43.0.1
slowapi==0.1.9